fix failing doctest
